fix(cloud): harden recurring reaper execution
