+++ server/.env.example
@@ -117,6 +117,7 @@ STRIPE_CUSTOMER_PORTAL_RETURN_URL=http://localhost:1420/settings/cloud
 # Sandbox provider
 E2B_API_KEY=
 E2B_TEMPLATE_NAME=
+CLOUD_SANDBOX_REAPER_GRACE_SECONDS=900.0
 E2B_WEBHOOK_SIGNATURE_SECRET=
 
 # Remote AnyHarness vendor Sentry injection into cloud sandboxes (hosted_product only in v1)
